share embedding & architectured tuned.

+++ src/connectfour/game.py
@@ -87,15 +87,15 @@ class BatchGameState(AbsBatchGameState):
 
     def _blank_board(self):
         return torch.tile(
-            torch.tensor(play_state_embedding(PlayState.BLANK))[:, None, None],
+            torch.tensor(play_state_embedding(PlayState.BLANK), dtype=torch.int32)[
+                :, None, None
+            ],
             (1, self._num_rows, self._num_cols),
         )
 
     def _blank_boards(self):
         blank_board = self._blank_board()
-        return torch.tile(
-            blank_board[None, :, :, :], (self.batch_size, 1, 1, 1)
-        )
+        return torch.tile(blank_board[None, :, :, :], (self.batch_size, 1, 1, 1))
 
     def winners(self, run_length=4) -> List[Optional[PlayState]]:
         results = np.array([None] * self.batch_size)
@@ -147,9 +147,7 @@ class BatchGameState(AbsBatchGameState):
         # Determine the number of previous plays in each column by summing the one hot mask:
         num_plays = torch.einsum(
             "ijki->i",
-            self._board_state[
-                :, play_state_embedding_ix(PlayState.BLANK) + 1 :, :, js
-            ],
+            self._board_state[:, play_state_embedding_ix(PlayState.BLANK) + 1 :, :, js],
         )
         is_ = self._num_rows - 1 - num_plays
         # Set the one-hot-values at those locations
@@ -192,9 +190,7 @@ class BatchGameState(AbsBatchGameState):
                         for row in game
                     ]
                 )
-                for game in torch.permute(
-                    self._board_state, (0, 2, 3, 1)
-                ).tolist()
+                for game in torch.permute(self._board_state, (0, 2, 3, 1)).tolist()
             ]
         )
 
@@ -202,9 +198,7 @@ class BatchGameState(AbsBatchGameState):
         game_strs = []
         for game in self.as_tuple():
             hor_line = "\n%s\n" % ("-" * (self._num_cols * 2 - 1))
-            game_strs.append(
-                hor_line.join(map(lambda row: "|".join(row), game))
-            )
+            game_strs.append(hor_line.join(map(lambda row: "|".join(row), game)))
         hor_line = "\n\n%s\n\n" % ("*" * (self._num_cols * 2 + 1))
         return hor_line.join(game_strs)
